cast the result to validated if the number is negative

+++ 0x1A-hash_tables/0-hash_table_create.c
@@ -11,7 +11,7 @@ hash_table_t *hash_table_create(unsigned long int size)
 	hash_table_t *hash_t = NULL;
 	unsigned long int i = 0;
 
-	if (size <= 0)
+	if ((int)size <= 0)
 		return (NULL);
 
 	hash_t =  malloc(sizeof(hash_table_t));
